src/simple/algos/authorizer.rs: Implement authorizer.
Sadly it seems that the authorizer needs to run completely in
any separated signer / secret-key-holder, due to the need to
select `k` by random.

This means that you need to modify any kind of "hardware wallet"
or similar to modify the signing algorithm to instead be a
DLEQ-making algorithm (signature algo is very similar to
DLEQ-making algo, and knowledge of the `k` behind a point
that is shown in `R` (for signature) / `d` (for our form of
DLEQ proof) is enough to leak secret key, so that all needs to
be done in the hardware wallet or separate secret-key-holder).